[WIP] Linker-level 32bit platform support

+++ javalib/src/main/scala/java/lang/Float.scala
@@ -241,7 +241,8 @@ object Float {
       errno.errno = 0
       val res = stdlib.strtof(cstr, end)
 
-      if (errno.errno == 0 && cstr != !end && string.strlen(!end) == 0) res
+      if (errno.errno == 0 && cstr != !end && string.strlen(!end).toInt == 0)
+        res
       else throw new NumberFormatException(s)
     }
